Remove non-material themes
