+++ src/app/@theme/components/header/header.component.html
@@ -7,9 +7,7 @@
       [matRippleUnbounded]="true"
       [matRippleCentered]="true"
       (click)="toggleSidebar()"
-    >
-      <nb-icon [icon]="(materialTheme$ | async) ? 'menu-outline' : 'menu-2-outline'"></nb-icon>
-    </a>
+    ><nb-icon icon="menu-outline"></nb-icon></a>
     <a class="logo" href="#" (click)="navigateHome()">ngx-<span>admin</span></a>
   </div>
   <nb-select
