+++ src/app/@theme/components/header/header.component.scss
@@ -19,6 +19,18 @@
     align-content: center;
   }
 
+  ::ng-deep {
+    nb-actions, .logo-container {
+      nb-icon {
+        color: nb-theme(color-basic-100) !important;
+      }
+    }
+
+    .select-button {
+      background-color: nb-theme(background-basic-color-3) !important;
+    }
+  }
+
   nb-user {
     cursor: pointer;
   }
