+++ src/app/@theme/styles/_angular-material.scss
@@ -1,0 +1,15 @@
+@import '~@angular/material/theming';
+
+@mixin angular-material() {
+  @include mat-core();
+
+  @include nb-for-theme(material-dark) {
+    $custom-dark-theme: mat-dark-theme(mat-palette($mat-pink), mat-palette($mat-blue-grey));
+    @include angular-material-theme($custom-dark-theme);
+  }
+  
+  @include nb-for-theme(material-light) {
+    $custom-light-theme: mat-light-theme(mat-palette($mat-indigo), mat-palette($mat-pink));
+    @include angular-material-theme($custom-light-theme);
+  }
+}
