+++ src/app/@theme/styles/_material-dark.scss
@@ -363,6 +363,8 @@ $theme: (
 
   font-family-primary: unquote('Roboto, sans-serif'),
 
+  layout-padding-top: 2.25rem,
+
   shadow: unquote('0px 2px 1px -1px rgba(0, 0, 0, 0.2), 0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 1px 3px 0px rgba(0, 0, 0, 0.12)'),
   card-shadow: shadow,
   header-shadow: unquote(
@@ -375,8 +377,6 @@ $theme: (
   footer-text-color: text-basic-color,
   footer-text-highlight-color: footer-text-color,
   sidebar-background-color: background-basic-color-2,
-  actions-icon-color: text-basic-color,
-  actions-text-color: text-basic-color,
 
   material-regular-font-weight: 400,
   menu-text-font-weight: material-regular-font-weight,
@@ -387,6 +387,7 @@ $theme: (
 
   menu-item-icon-color: rgba(255, 255, 255, 0.7),
   menu-item-icon-hover-color: rgba(255, 255, 255, 0.7),
+  menu-item-icon-margin: 0 0.5rem 0 0,
 
   menu-submenu-item-hover-background-color: rgba(255, 255, 255, 0.04),
   menu-submenu-item-active-hover-background-color: rgba(255, 255, 255, 0.1),
@@ -395,6 +396,20 @@ $theme: (
   card-border-style: none,
   card-background-color: color-basic-800,
   card-divider-color: color-basic-700,
+  card-height-tiny: 13.5rem,
+  card-height-small: 21.1875rem,
+  card-height-medium: 28.875rem,
+  card-height-large: 36.5625rem,
+  card-height-giant: 44.25rem,
+  card-margin-bottom: 1.875rem,
+  card-header-with-select-padding-top: 0.5rem,
+  card-header-with-select-padding-bottom: 0.5rem,
+
+  select-min-width: 6rem,
+
+  slide-out-background: linear-gradient(270deg, #1f1f1f 0%, #292929 100%),
+  slide-out-shadow-color: 0 4px 14px 0 #292929,
+  slide-out-shadow-color-rtl: 0 4px 14px 0 #292929,
 
   input-border-width: 0 0 1px 0,
   input-basic-border-color: rgba(255, 255, 255, 0.7),
@@ -493,4 +508,4 @@ $theme: (
   option-giant-text-font-weight: material-regular-font-weight
 );
 
-$nb-themes: nb-register-theme($theme, material-dark, dark);
+$nb-themes: nb-register-theme($theme, material-dark, default);
