+++ src/app/@theme/styles/_overrides.scss
@@ -12,4 +12,33 @@
   .table {
     color: nb-theme(text-basic-color) !important;
   }
+
+  nb-sidebar {
+    transition: width 0.3s;
+
+    .main-container {
+      transition: width 0.3s;
+    }
+  }
+
+  nb-card {
+    border-bottom-left-radius: 0.125rem;
+    border-bottom-right-radius: 0.125rem;
+
+    nb-card-header, .tabset {
+      background-color: nb-theme(card-divider-color);
+    }
+  }
+
+  [nbinput] {
+    font-weight: 400;
+
+    &.status-basic:focus:hover {
+      border-color: nb-theme(color-primary-focus) !important;
+    }
+  }
+
+  [nbbutton] {
+    box-shadow: none !important;
+  }
 }
