--- src/app/@theme/styles/material/_material.scss
@@ -1,57 +1,0 @@
-@import '~@angular/material/theming';
-
-@include nb-install() {
-  @include nb-for-theme(material-dark) {
-    $custom-dark-theme: mat-dark-theme(mat-palette($mat-pink), mat-palette($mat-blue-grey));
-    @include angular-material-theme($custom-dark-theme);
-  }
-
-  @include nb-for-theme(material-light) {
-    $custom-light-theme: mat-light-theme(mat-palette($mat-indigo), mat-palette($mat-pink));
-    @include angular-material-theme($custom-light-theme);
-  }
-
-  @include nb-for-themes(material-dark, material-light) {
-    @include mat-core();
-
-    nb-layout-header {
-      .logo-container nb-icon {
-        color: nb-theme(text-control-color);
-      }
-
-      .select-button {
-        background-color: nb-theme(background-basic-color-3) !important;
-      }
-    }
-    
-
-    nb-sidebar {
-      transition: width 0.3s;
-
-      .main-container {
-        transition: width 0.3s;
-      }
-    }
-
-    nb-card {
-      border-bottom-left-radius: 0.125rem;
-      border-bottom-right-radius: 0.125rem;
-
-      nb-card-header, .tabset {
-        background-color: nb-theme(card-divider-color);
-      }
-    }
-
-    [nbinput] {
-      font-weight: 400;
-
-      &.status-basic:focus:hover {
-        border-color: nb-theme(color-primary-focus) !important;
-      }
-    }
-
-    [nbbutton] {
-      box-shadow: none !important;
-    }
-  }
-}
