+++ src/app/@theme/theme.module.ts
@@ -35,12 +35,8 @@ import {
   ThreeColumnsLayoutComponent,
   TwoColumnsLayoutComponent,
 } from './layouts';
-import { DEFAULT_THEME } from './styles/theme.default';
-import { COSMIC_THEME } from './styles/theme.cosmic';
-import { CORPORATE_THEME } from './styles/theme.corporate';
-import { DARK_THEME } from './styles/theme.dark';
-import { MATERIAL_LIGHT_THEME } from './styles/material/theme.material-light';
-import { MATERIAL_DARK_THEME } from './styles/material/theme.material-dark';
+import { MATERIAL_LIGHT_THEME } from './styles/theme.material-light';
+import { MATERIAL_DARK_THEME } from './styles/theme.material-dark';
 
 const NB_MODULES = [
   NbLayoutModule,
@@ -84,10 +80,8 @@ export class ThemeModule {
       ngModule: ThemeModule,
       providers: [
         ...NbThemeModule.forRoot(
-          {
-            name: 'default',
-          },
-          [ DEFAULT_THEME, COSMIC_THEME, CORPORATE_THEME, DARK_THEME, MATERIAL_LIGHT_THEME, MATERIAL_DARK_THEME ],
+          { name: 'material-light' },
+          [MATERIAL_LIGHT_THEME, MATERIAL_DARK_THEME],
         ).providers,
       ],
     };
