+++ src/app/pages/dashboard/rooms/room-selector/room-selector.component.ts
@@ -1,19 +1,12 @@
-import { Component, EventEmitter, HostBinding, OnDestroy, OnInit, Output } from '@angular/core';
+import { Component, EventEmitter, Output } from '@angular/core';
 import { Location, LocationStrategy } from '@angular/common';
-import { NbThemeService } from '@nebular/theme';
-import { map, takeUntil } from 'rxjs/operators';
-import { Subject } from 'rxjs';
 
 @Component({
   selector: 'ngx-room-selector',
   templateUrl: './room-selector.component.html',
   styleUrls: ['./room-selector.component.scss'],
 })
-export class RoomSelectorComponent implements OnInit, OnDestroy {
-
-  private destroy$ = new Subject<void>();
-  private hideGrid: boolean;
-
+export class RoomSelectorComponent {
   @Output() select: EventEmitter<number> = new EventEmitter();
 
   selectedRoom = null;
@@ -66,33 +59,8 @@ export class RoomSelectorComponent implements OnInit, OnDestroy {
     ],
   };
 
-  @HostBinding('style.background')
-  get background(): 'none' | null {
-    return this.hideGrid ? 'none' : null;
-  }
-
-  constructor(
-    private location: Location,
-    private locationStrategy: LocationStrategy,
-    private themeService: NbThemeService,
-  ) {
+  constructor(private location: Location, private locationStrategy: LocationStrategy) {
     this.selectRoom('2');
-  }
-
-  ngOnInit() {
-    this.hideGrid = this.themeService.currentTheme === 'corporate';
-
-    this.themeService.onThemeChange()
-      .pipe(
-        map(({ name }) => name === 'corporate'),
-        takeUntil(this.destroy$),
-      )
-      .subscribe((hideGrid: boolean) => this.hideGrid = hideGrid);
-  }
-
-  ngOnDestroy() {
-    this.destroy$.next();
-    this.destroy$.complete();
   }
 
   private sortRooms() {
